easy-rsa: Unify KEY_CN, KEY_NAME and FN setup in pkitool
Making it special case for root CA does not look right. It seems the
only purpose for this is to set KEY_CN to "$KEY_ORG CA" for root CA
when KEY_CN is empty or unset and set FN to "ca" statically.

On the other hand KEY_NAME is set to FN with stripped space using tr(1)
that in case of root CA always "ca" ignoring "$1" non-option argument
of pkitool for FN/KEY_NAME.

Unify root CA and regular certificate use cases, set FN statically to
"ca" for root CA case.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>